First 5 tasks completed.
TASK-1.2.1: Prime number function: use while instead of for
TASK-1.2.2: First 100 prime numbers: use for instead of while
TASK-W1-3A: Write a piece of code that will generate numbers for a coupon in Sayisal Loto that should include 8 colons and each column should have 6 random numbers. Please control the colons in terms of having no repetitions.
TASK-W1-3B: Please write a piece of code to find the names and order no of the presidents with the highest and the lowest heights.
Task-3A: Please provide a similar table for your favorite 7 cities of Turkey that include their area and population.